Disable post button when there is no content
## Background
The current version of the app allows you to post even if you haven't typed in any content or uploaded any images. A post isn't actually created on micro.blog.

## Modifications
I created a stored property for the `ImageState` on the `ComposeViewModel`. We need to store this so we can evaluate if the post button should be enabled/disabled based on text length, image uploading status, and also how many images have been uploaded. Please let me know if you think there was a better way to do this.

## Testing
* By default the post button should be disabled when creating a new post.
* If you upload an image the post button should be enabled after it is uploaded.
* If you type in any text it should be come enabled. If you delete the text and don't have any images uploaded, it should be disabled.
* If are uploading an image you should be able to type text. The post button shouldn't be enabled because the image is still uploading.